requirements.txt
streamlit
openai
langchain
langchain-community
python-dotenv
pandas
chardet
sentence-transformers 